requirements.txt

pydig
pyyaml
requests
dateparser
tldextract
python-whois
